#7 - Create register handler (#9)
* #7 - Create register handler
If the event is coming from /register then we execute register handler.
Note we do not need to check if it is a POST or not, because register
route is already defined in API Gateway as post.

Write permissions were added to the role of Lambda in order to give it
access to DynamoDB.The need to automate the infra is becoming obvious.
There are parts infra parts of which to keep a mental account.

* #7 Access tablename from env
Access the users table name from repository enviroment variable.Also
removed the DynamoDB access key and secret.They are not needed.

* #7 Add put item condition
Added condition so that only users with unique emails are registered.
Setting the email as partition key, i.e. primary key, was not sufficient
to ensure email uniqueness.

Also added proper HTTP codes and messages for the different errors.