#6 create login handler (#10)
* #6 Create login handler
On login compare the passwords of user. This commit also includes
a bit of refactoring.The register and login handler share a lot of
code.

* #6 Return token
Return a jwt token with an expiry date of 3 days. Currently
the payload is quite bare, in fact contains only email. As the user
schema evolves so will the payload.