Clean up welcome email dispatch
Removed token retrieval from oban job handler and separated email
delivery from email setup.The idea is to avoid side effects, as much as
possib, in the functions.This makes them easy to reason about. This
change also gives us a nice separation of concern.

A tag was also added to the new_user job, this allows to match on the
tag in the perform function.This change essentially makes the flow very
much like redux, where new_user dispatches an action  where the tag acts
as action type and perform  acts as a reducer that updates the state
according what the tag is.