Add verify function to accounts
Accessing all users/accounts via the accounts context abstracts the db
information from the access layer. For example, if users schema changes
so that "verified" field has changed to "confirmed" then access layer
does not need to be updated, the update is localized to the accounts
context.

The more we add to the accounts context the more the need to refactor
the queries becomes necessary.