Clean up user controller
The changes from the previous commits are taken into account, and any
code that was matching on the User module are removed.The status of the
accounts operation are enough.Important to mention again the db schema
is abstracted from the access layer,i.e. controller, so no need to know
what the shape of the data returned is.

For example, {:ok, user} <- Accounts.create_user(...) is much better
than {:ok, %User{}} <- Account.create_user(...).