Test for post forgot password
A 200 response is sent back whethere there is user account for the given
email or not. In case there is no record a 200 is sent in order not to
leak information.

The tests also the queuing of oban jobs, of course this is not ideal
because at the controller level the structure of the job is abstracted,
but this stop gap solution until mocking modules is added to the tests.