Set default value of verified field
Set the default value of verified field to false.This bring the Ecto
schema and db schema to parity.This was necessary, because when
inserting a user the struct returned by Ecto had a null verified while
in the database the field was false.

Note that it is probably the right time to separate Account from
User.