40 add endpoint for forgotten password (#43)
* Clean up welcome email dispatch

Removed token retrieval from oban job handler and separated email
delivery from email setup.The idea is to avoid side effects, as much as
possib, in the functions.This makes them easy to reason about. This
change also gives us a nice separation of concern.

A tag was also added to the new_user job, this allows to match on the
tag in the perform function.This change essentially makes the flow very
much like redux, where new_user dispatches an action  where the tag acts
as action type and perform  acts as a reducer that updates the state
according what the tag is.

* Collected auth related code under auth namespace

With auth related code in one namespace, we get a nice separation of
concerns and the auth implementation details are abstracted away from
other modules.

Not sure if placing the auth in the access layer, i.e. UsersWeb
,namespace is correct decision, but it is easy to move it elsewhere if
necessary.

* Add verify function to accounts

Accessing all users/accounts via the accounts context abstracts the db
information from the access layer. For example, if users schema changes
so that "verified" field has changed to "confirmed" then access layer
does not need to be updated, the update is localized to the accounts
context.

The more we add to the accounts context the more the need to refactor
the queries becomes necessary.

* Clean up user controller

The changes from the previous commits are taken into account, and any
code that was matching on the User module are removed.The status of the
accounts operation are enough.Important to mention again the db schema
is abstracted from the access layer,i.e. controller, so no need to know
what the shape of the data returned is.

For example, {:ok, user} <- Accounts.create_user(...) is much better
than {:ok, %User{}} <- Account.create_user(...).

* Remove dead code and update auth module references.

* Test for post forgot password

A 200 response is sent back whethere there is user account for the given
email or not. In case there is no record a 200 is sent in order not to
leak information.

The tests also the queuing of oban jobs, of course this is not ideal
because at the controller level the structure of the job is abstracted,
but this stop gap solution until mocking modules is added to the tests.

* Add forgot password logic

When a forgot password request is received, an email with a reset
password url is sent to the user. Note the similarity to the signup
flow.

* Make oban testing manual.

* Update config for auth module.

* Rename events to worker.

* Add reset password url to configs.

* Clean up tests.

* Add reset password url to ci so that it is available at build time.

* Set default value of verified field

Set the default value of verified field to false.This bring the Ecto
schema and db schema to parity.This was necessary, because when
inserting a user the struct returned by Ecto had a null verified while
in the database the field was false.

Note that it is probably the right time to separate Account from
User.

* Reinstate github action job trigger rules.