Convert to umbrella app
Separate the access layer, users_web and core layer, users_core into two
separate applications. The users_web is app is dependent on users_core
so users_core is started before users_web, see the :users_core in
users_web mix file.

This conversion is done to move the application to a hexagonal-esque
architecture, and in doing so make it easier to add more access layers.