Separate test deps
The test dependecies mox and ex_machina are NOT in fact needed for both
apps. Mox is for mocking the http client, so needed only in users_web
whereas ex_machina is for seeding database with mock data so only needed
in users_core app.

The only deps that is common in credo which is used for static
analysis.Having at the top level allows us to run `mix credo` there.