46 convert to umbrella app (#47)
* Convert to umbrella app

Separate the access layer, users_web and core layer, users_core into two
separate applications. The users_web is app is dependent on users_core
so users_core is started before users_web, see the :users_core in
users_web mix file.

This conversion is done to move the application to a hexagonal-esque
architecture, and in doing so make it easier to add more access layers.

* Add hackney dependency

Add hackney dependency that is needed by swoosh.Previously, it seemed to
have been installed locally which updated the mix.lock file, the result
being every time mix deps.get task was run it was installed
successfully.Only after removing the mix.lock file and trying to compile
that app was this oversight caught.

* Use mix aliases

Use the mix alias to run both db creation and migration in init script.
Apparently, mix aliases are exactly like npm scripts.

* Update readmes

The `mix phx.new ... --umbrella` helpfully creates Readmes at each
level, so updating them with a quib about each app.

* Update mix lock file.

* Refactor deps list.

* 46 Run tasks in users-umbrella folder.

* Move common deps to umbrella level.

* Separate test deps

The test dependecies mox and ex_machina are NOT in fact needed for both
apps. Mox is for mocking the http client, so needed only in users_web
whereas ex_machina is for seeding database with mock data so only needed
in users_core app.

The only deps that is common in credo which is used for static
analysis.Having at the top level allows us to run `mix credo` there.